requirements.txt
requests~=2.32.3
flask~=3.0.2
openai~=1.16.1
gunicorn==21.2.0
python-dotenv~=1.0.1
paho-mqtt~=2.1.0